draw old using

+++ src/utils.py
@@ -25,7 +25,7 @@ def solve_equation(equations, solutions):
             spEqs.append(sympy_eq)   
         solved = solve(spEqs, dict=True)
         # cur_targets = [round(i) for i in list(solved[0].values())]
-        act_solns = [round(i) for i in (list(solved[0].values()))]
+        act_solns = [int(round(i)) for i in (list(solved[0].values()))]
 
         print("act", act_solns)
         print("pred", solutions)
